requirements.txt
google-adk>=0.1.0
google-genai>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
deprecated>=1.2.14
